+++ src/StatePulse.NET/Engine/Implementations/StatePulseRegistry.cs
@@ -2,21 +2,21 @@
 internal class StatePulseRegistry : IStatePulseRegistry
 {
     private readonly List<Type> _knownStates = new();
-    private readonly List<Type> _knownEffects = new();
-    private readonly List<Type> _knownReducers = new();
+    private readonly Dictionary<Type, Type> _knownEffects = new();
+    private readonly Dictionary<Type, Type> _knownReducers = new();
     private readonly List<Type> _knownActions = new();
-    private readonly List<Type> _knownActionValidators = new();
+    private readonly Dictionary<Type, Type> _knownActionValidators = new();
     //TODO: SWITCH TO ConcurrentStack perform as a list avoid duplicate action as key...
 
     public IReadOnlyList<Type> KnownStates => _knownStates;
-    public IReadOnlyList<Type> KnownEffects => _knownEffects;
-    public IReadOnlyList<Type> KnownReducers => _knownReducers;
+    public IReadOnlyDictionary<Type, Type> KnownEffects => _knownEffects;
+    public IReadOnlyDictionary<Type, Type> KnownReducers => _knownReducers;
     public IReadOnlyList<Type> KnownActions => _knownActions;
-    public IReadOnlyList<Type> KnownActionValidators => _knownActionValidators;
+    public IReadOnlyDictionary<Type, Type> KnownActionValidators => _knownActionValidators;
 
-    public void RegisterEffect(Type effectType) => _knownEffects.Add(effectType);
-    public void RegisterReducer(Type reducerType) => _knownReducers.Add(reducerType);
+    public void RegisterEffect(Type effectType, Type interfaceType) => _knownEffects.Add(effectType, interfaceType);
+    public void RegisterReducer(Type reducerType, Type interfaceType) => _knownReducers.Add(reducerType, interfaceType);
     public void RegisterState(Type stateType) => _knownStates.Add(stateType);
     public void RegisterAction(Type actionType) => _knownActions.Add(actionType);
-    public void RegisterActionValidator(Type actionValType) => _knownActionValidators.Add(actionValType);
+    public void RegisterActionValidator(Type actionValType, Type interfaceType) => _knownActionValidators.Add(actionValType, interfaceType);
 }