+++ src/StatePulse.NET/IActionValidator.cs
@@ -1,10 +1,7 @@
 ﻿namespace StatePulse.Net;
-/// <summary>
-/// Use to define validation behavior for an action... <br/>
-/// some business logic are required under certain rules and those rules can be defined here! if the validation fails the action is not dispatched
-/// </summary>
-/// <typeparam name="TAction"></typeparam>
-public interface IActionValidator<in TAction> where TAction : IAction
+public interface IActionValidator<in TAction, TEffect>
+    where TAction : IAction
+    where TEffect : IEffect<TAction>
 {
-    void Validate(TAction action, ref ValidationResult result);
+    Task<bool> Validate(TAction action);
 }