+++ src/StatePulse.NET/IStatePulseRegistry.cs
@@ -3,14 +3,14 @@ public interface IStatePulseRegistry
 {
 
     IReadOnlyList<Type> KnownStates { get; }
-    IReadOnlyList<Type> KnownEffects { get; }
-    IReadOnlyList<Type> KnownReducers { get; }
+    IReadOnlyDictionary<Type, Type> KnownEffects { get; }
+    IReadOnlyDictionary<Type, Type> KnownReducers { get; }
     public IReadOnlyList<Type> KnownActions { get; }
-    public IReadOnlyList<Type> KnownActionValidators { get; }
+    public IReadOnlyDictionary<Type, Type> KnownActionValidators { get; }
     void RegisterState(Type stateType);
-    void RegisterEffect(Type effectType);
-    void RegisterReducer(Type reducerType);
+    void RegisterEffect(Type effectType, Type interfaceType);
+    void RegisterReducer(Type reducerType, Type interfaceType);
     void RegisterAction(Type actionType);
-    void RegisterActionValidator(Type actionValType);
+    void RegisterActionValidator(Type actionValType, Type interfaceType);
 
 }